add import for test; refactor mdgopackm to packm
