+++ tests/test_mdgopackmol.py
@@ -7,7 +7,7 @@ import pytest
 import numpy as np
 from pymatgen.core import Molecule
 
-from mdgo.util.mdgopackmol import PackmolWrapper
+from mdgo.util.packmol import PackmolWrapper
 
 test_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "test_files")
